+++ CHANGELOG.md
@@ -9,6 +9,7 @@
    * Fixed case-sensitive `accept` in `filepicker` field
    * Fixed pages admin being accessible without read/write permissions [grav-plugin-admin#2053](https://github.com/getgrav/grav-plugin-admin/issues/2053)
    * Fixed missing event `onAdminCreatePageFrontmatter` when creating a new page [grav-plugin-auto-date#8](https://github.com/getgrav/grav-plugin-auto-date/issues/8)
+   * Fixed missing event `onAdminAfterDelMedia` when deleting a file from a page
 
 # v1.0.2
 ## 02/01/2021